ClientApp query input handled
The client app can handle every type of query provided in the user input (doesn't fully check if the command has any error). Also implemented a loop that checks if there is any "Result" (server response) to be read, and if not it waits for the user's input. Function handleResult() is still under development